Re-written directory portion to fix bugs, clean up logic, and greatly increase efficiency.
Instead of pulling all nodes from the database and then filtering them afterwards, we now intelligently query the database so that all nodes retrieved from the database are returned as search results (none need to be thrown out). This fixes the bugs resulting from our cap on the maximum search size preventing us from finding the correct node for a search if it was out of scope.

Upgraded Apache Mina jar to the latest stable version (previously we used a release candidate).

Ported vLdap from using the ApacheDS jars that we needed to extract from the zip file every build to using the new Apache LDAP API project which has been created specifically for the purpose of allowing third parties to include LDAP functionality in their applications. Now we only need to include one jar instead of over ten jars. This will also, hopefully, allow us to upgrade simply by dropping in the latest jar from that project.

Because of moving to the new LDAP jar I was able to find a way to completely remove both the schema and the schema manager from vLdap. This simplified the code immensely, and increases vLdap's startup speed by orders of magnitude because it no longer needs to load the schema each time it starts. This also means we don't need to ship the LDAP schema anymore.

Added jars

Added new files

Removing obsolete zip

Source formatting

Source formatting

Source formatting