LPS-49798 add column to store the original URL of an article
For articles whose authoritative origin isn't KB itself, store the
source location in this new column. The value of this source location is
expected to be any valid URL.

This can be used to identify and offer extended functionality for
articles imported from Github, Bitbucket, etc. For the moment, only
support for GitHub will be implemented.